chore(performance): speed up csm detection[]
