+++ package.json
@@ -22,6 +22,7 @@
     "lint": "lerna run lint",
     "tsc": "lerna run tsc",
     "test": "lerna run --stream test",
+    "benchmark": "lerna run benchmark",
     "test:ci": "lerna run test:ci",
     "coverage": "lerna run coverage",
     "version": "lerna version --no-private --conventional-commits --create-release github --yes --exact",
