+++ packages/live-preview-sdk/src/inspectorMode/__tests__/getAllTaggedElements.test.ts
@@ -1,42 +1,15 @@
-import { SourceMapMetadata, combine } from '@contentful/content-source-maps';
+import { combine } from '@contentful/content-source-maps';
 import { describe, expect, it } from 'vitest';
 
 import { InspectorModeDataAttributes } from '../types.js';
 import { getAllTaggedElements } from '../utils.js';
+import { createSourceMapFixture } from './fixtures/contentSourceMap.js';
 
 describe('getAllTaggedElements', () => {
   const dataEntry = InspectorModeDataAttributes.ENTRY_ID;
   const dataAsset = InspectorModeDataAttributes.ASSET_ID;
   const dataField = InspectorModeDataAttributes.FIELD_ID;
   const dataLocale = InspectorModeDataAttributes.LOCALE;
-
-  function createSourceMapFixture(
-    entityId: string,
-    overrides?: {
-      origin?: string;
-      contentful?: Partial<Omit<SourceMapMetadata['contentful'], 'entity'>>;
-    },
-  ): SourceMapMetadata {
-    const origin = overrides?.origin ?? 'contentful.com';
-    const space = overrides?.contentful?.space ?? 'master';
-    const environment = overrides?.contentful?.environment ?? 'master';
-    const entityType = overrides?.contentful?.entityType ?? 'Entry';
-    const locale = overrides?.contentful?.locale ?? 'en-US';
-    const field = overrides?.contentful?.field ?? 'title';
-
-    return {
-      href: `https://app.${origin}/spaces/${space}/environments/${environment}/entries/${entityId}?focusedField=${field}&focusedLocale=${locale}`,
-      origin,
-      contentful: {
-        space,
-        environment,
-        entity: entityId,
-        entityType,
-        field,
-        locale,
-      },
-    };
-  }
 
   const html = (text: string) => {
     return new DOMParser().parseFromString(text, 'text/html');
@@ -279,94 +252,6 @@ describe('getAllTaggedElements', () => {
       );
       expect(elements[0].getAttribute(InspectorModeDataAttributes.FIELD_ID)).toEqual('title');
       expect(elements[0].getAttribute(InspectorModeDataAttributes.LOCALE)).toEqual('en-US');
-    });
-
-    it('works performant on larger pages', () => {
-      const dom = html(`
-      <div id="root">
-        <div id="parent_tagged">
-          ${new Array(10)
-            .fill('x')
-            .map((_, index) => {
-              const sourceMap = createSourceMapFixture('parent_tagged');
-              const content = combine(
-                `Parent should be tagged even if this is the "${index}" element.`,
-                sourceMap,
-              );
-              return `<p>${content}</p>`;
-            })
-            .join('')}
-        </div>
-        <div id="flat_list">
-          ${new Array(10)
-            .fill('x')
-            .map((_, index) => {
-              const sourceMap = createSourceMapFixture(`flat_item-${index}`);
-              const content = combine(
-                `Every item in the flat list should be tagged. This is the ${index} element.`,
-                sourceMap,
-              );
-              return `<p>${content}</p>`;
-            })
-            .join('')}
-        </div>
-        <div id="richtext">
-          ${new Array(30)
-            .fill('x')
-            .map((_, index) => {
-              const sourceMap = createSourceMapFixture(`rte-nested-${index}`);
-              const inlineItems = new Array(10).fill('x').map((_, nestedIndex) => {
-                const tag = nestedIndex % 2 === 0 ? 'strong' : 'b';
-                const content = combine(
-                  `Richtext nested content with tag ${tag} ${index} ${nestedIndex}`,
-                  sourceMap,
-                );
-                return `<${tag}>${content}</${tag}>`;
-              });
-
-              const inlineContent = inlineItems.join(' ');
-
-              return `<p id="richtext-paragraph-${index}">${inlineContent}</p>`;
-            })
-            .join('\n')}
-        </div>
-        <div id="img">
-          <img
-            src="./picture.jpg"
-            alt="${combine('Some alt text for the picture', createSourceMapFixture('img-1', { contentful: { entityType: 'Asset' } }))}"
-          />
-        </div>
-        <div id="picture">
-          <picture>
-            <source srcset="/lion-potrait.jpg" media="(orientation: portrait)" />
-            <img
-              src="/lion-298-332.jpg"
-              alt="${combine('A lion staring at the sun', createSourceMapFixture('img-2', { contentful: { entityType: 'Asset' } }))}"
-            />
-          </picture>
-        </div>
-        <div id="figure">
-          <figure>
-            <span>
-              <span>
-                <img
-                  src="/elephant.jpg"
-                  alt="${combine('Elephant at sunset', createSourceMapFixture('img-3', { contentful: { entityType: 'Asset' } }))})}"
-                />
-              </span>
-            </span>
-            <figcaption>${combine('An elephant at sunset', createSourceMapFixture('img-3', { contentful: { entityType: 'Asset' } }))})}</figcaption>
-          </figure>
-        </div>
-      </div>
-    `);
-
-      const starTime = performance.now();
-      const elements = getAllTaggedElements(dom, true);
-      const diff = performance.now() - starTime;
-
-      expect(elements).toHaveLength(1 + 10 + 30 + 3);
-      expect(diff).toBeLessThanOrEqual(100);
     });
   });
 });
